requirements.txt
blinker==1.7.0
click==8.1.7
Deprecated==1.2.14
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysqlclient==2.2.0
ordered-set==4.1.0
packaging==23.2
Pygments==2.16.1
python-dotenv==1.0.0
rich==13.6.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.1
